merge code from postgres/postgres

+++ src/include/access/nbtree.h
@@ -168,7 +168,7 @@ typedef struct BTMetaPageData
 /*
  * MaxTIDsPerBTreePage is an upper bound on the number of heap TIDs tuples
  * that may be stored on a btree leaf page.  It is used to size the
- * per-page temporary buffers used by index scans.)
+ * per-page temporary buffers used by index scans.
  *
  * Note: we don't bother considering per-tuple overheads here to keep
  * things simple (value is based on how many elements a single array of
@@ -381,7 +381,7 @@ static inline void
 BTreeTupleSetPosting(IndexTuple itup, int nhtids, int postingoffset)
 {
 	Assert(nhtids > 1 && (nhtids & BT_OFFSET_MASK) == nhtids);
-	Assert(postingoffset == MAXALIGN(postingoffset));
+	Assert((size_t) postingoffset == MAXALIGN(postingoffset));
 	Assert(postingoffset < INDEX_SIZE_MASK);
 
 	itup->t_info |= INDEX_ALT_TID_MASK;
